move functions
move _compute_type_chart() to data.py
move raw stats conversion to utils.py
