+++ src/poke_env/teambuilder/teambuilder_pokemon.py
@@ -2,20 +2,12 @@
 """This module defines the TeambuilderPokemon class, which is used as an intermediate
 format to specify pokemon builds in teambuilders custom classes.
 """
+import poke_env.utils
 from poke_env.utils import to_id_str
 
 
 class TeambuilderPokemon:
-    STATS_TO_IDX = {
-        "hp": 0,
-        "atk": 1,
-        "def": 2,
-        "spa": 3,
-        "spd": 4,
-        "spe": 5,
-        "satk": 3,
-        "sdef": 4,
-    }
+    STATS_TO_IDX = poke_env.utils.STATS_TO_IDX
 
     HP_TO_IVS = {
         "bug": [31, 31, 31, 30, 31, 30],
